Refinamento dos domínios, criação de classes de exceções personalizadas e melhoras nas DTOs de avaliações.

+++ src/main/java/com/example/demo/controller/dto/MediaPerformanceDTO.java
@@ -1,8 +1,10 @@
 package com.example.demo.controller.dto;
 
+import java.math.BigDecimal;
+
 public record MediaPerformanceDTO(
-        double mediaComportamental,
-        double mediaEntregas,
-        double notaFinal
+        BigDecimal mediaComportamental,
+        BigDecimal mediaEntregas,
+        BigDecimal notaFinal
 ) {
 }
